backend: Hide menu when rebuilding menu items
Since we have a monitor to update the menu whenever the desktop files
are updated, the list of filtered menu items can change while the user
is actively navigating through it. This causes the GtkListBox filters to
segfault.

Hiding the menu whenever the GAppInfoMonitor fires the "changed" signal
clears out the GtkListBox, and so no filtering is happening. This only
happens when installing or removing applications while browsing the
menu, so it should not be generally disruptive, and will prevent the
menu from crashing.